update: record protocol treasury balances
